Added the configuration for case study 2a
The diffrence of the configuration for the case study 2a
from the one for the case study 1a is
* Nhrz(length of the horizon) <- (1, 4, 16),
Nhrz is selected among the options, randomly.

All the hyperparameters other than Nhrz are the same as follows:
* agentClass = "agent001" (x(t+1) = A * x(t) + B * u(t))
* NhiddenAgent <- (2, 4, 8)
* NhiddenEnv = NhiddenAgent ( assume that the number of poles are known)
* seed @ env <- (0, 1, 2)

It's expected that
as Nhrz increases,
the discrepancy of the pole distribution get improved.